+++ src/commonMain/kotlin/csense/kotlin/annotations/numbers/DoubleLimit.kt
@@ -2,7 +2,6 @@
 
 package csense.kotlin.annotations.numbers
 
-//region Double limit
 @Target(
     AnnotationTarget.FUNCTION,
     AnnotationTarget.PROPERTY,
@@ -16,48 +15,7 @@ package csense.kotlin.annotations.numbers
 )
 @Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
 @MustBeDocumented
-annotation class DoubleLimit(
-    val from: Double,
-    val to: Double
+public annotation class DoubleLimit(
+    val from: Double = Double.MIN_VALUE,
+    val to: Double = Double.MAX_VALUE
 )
-//endregion
-
-//region Double limit to max
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER,
-    AnnotationTarget.VALUE_PARAMETER,
-    AnnotationTarget.FIELD,
-    AnnotationTarget.LOCAL_VARIABLE,
-    AnnotationTarget.ANNOTATION_CLASS,
-    AnnotationTarget.TYPE//until you can annotate functional parameters
-)
-
-@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
-@MustBeDocumented
-annotation class DoubleLimitToMax(
-    val from: Double
-)
-//endregion
-
-//region Double limit from min
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER,
-    AnnotationTarget.VALUE_PARAMETER,
-    AnnotationTarget.FIELD,
-    AnnotationTarget.LOCAL_VARIABLE,
-    AnnotationTarget.ANNOTATION_CLASS,
-    AnnotationTarget.TYPE//until you can annotate functional parameters
-)
-
-@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
-@MustBeDocumented
-annotation class DoubleLimitFromMin(
-    val from: Double
-)
-//endregion