+++ src/commonMain/kotlin/csense/kotlin/annotations/numbers/FloatLimit.kt
@@ -2,7 +2,6 @@
 
 package csense.kotlin.annotations.numbers
 
-//region float limit
 @Target(
     AnnotationTarget.FUNCTION,
     AnnotationTarget.PROPERTY,
@@ -16,47 +15,7 @@ package csense.kotlin.annotations.numbers
 
 @Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
 @MustBeDocumented
-annotation class FloatLimit(
-    val from: Float,
-    val to: Float
+public annotation class FloatLimit(
+    val from: Float = Float.MIN_VALUE,
+    val to: Float = Float.MAX_VALUE
 )
-//endregion
-
-//region float limit to max
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER,
-    AnnotationTarget.VALUE_PARAMETER,
-    AnnotationTarget.FIELD,
-    AnnotationTarget.LOCAL_VARIABLE,
-    AnnotationTarget.ANNOTATION_CLASS
-)
-
-@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
-@MustBeDocumented
-annotation class FloatLimitToMax(
-    val from: Float
-)
-//endregion
-
-//region float limit from min
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER,
-    AnnotationTarget.VALUE_PARAMETER,
-    AnnotationTarget.FIELD,
-    AnnotationTarget.LOCAL_VARIABLE,
-    AnnotationTarget.ANNOTATION_CLASS,
-    AnnotationTarget.TYPE//until you can annotate functional parameters
-)
-
-@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
-@MustBeDocumented
-annotation class FloatLimitFromMin(
-    val from: Float
-)
-//endregion