+++ src/commonMain/kotlin/csense/kotlin/annotations/numbers/LongLimit.kt
@@ -16,7 +16,7 @@ package csense.kotlin.annotations.numbers
 
 @Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
 @MustBeDocumented
-annotation class LongLimit(
+public annotation class LongLimit(
     val from: Long = Long.MIN_VALUE,
     val to: Long = Long.MAX_VALUE
 )