+++ src/commonMain/kotlin/csense/kotlin/annotations/threading/InUi.kt
@@ -11,4 +11,4 @@ package csense.kotlin.annotations.threading
 )
 @Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
 @MustBeDocumented
-annotation class InUi
+public annotation class InUi