+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.0.19
+- use kotlin 1.4m2 (useful for double& float Limit)
+    - {Double, Float}Limit have only 1 annotation now (with defaults)
+
+- UnitSystem annotation added
+- kotlin explicit library mode enabled & enforced
 # 0.0.18
 - added AnnotationTargets.Type as a workaround for the inability to annotate functional parameters.